LadderGame 클래스에서 GameMap 생성하던 로직을 GameMap 클래스로 이동
맵 초기화 책임을 LadderGame 보다 GameMap 이 갖고있느것이 더 적절하다고 판단하여,
생성 전략(로직)을 GameMap 으로 이동하였습니다.
